requirements.txt
pandas
plotly
streamlit
requests
python-dotenv
openai
pypdf
python-docx
